Initial commit of SphereMotionTeapot for Android
Basic structure of the teapot app. The app will throw up the startup
activity and, once connected, begin streaming the robot data. At this
point, only a log message is displayed when data from the robot is
received. Since this is being developed along with the RobotLibrary
commands and responses, the next step will be to extract the proper
data from the async data responses once the actual responses are built
out in the RobotLibrary.